+++ packages/tswebextension/src/lib/mv3/background/engine-api.ts
@@ -17,7 +17,7 @@ import browser from 'webextension-polyfill';
 import { type IFilter } from '@adguard/tsurlfilter/es/declarative-converter';
 import { type AnyRule } from '@adguard/agtree';
 
-import { QUICK_FIXES_FILTER_ID, USER_FILTER_ID } from '../../common/constants';
+import { USER_FILTER_ID } from '../../common/constants';
 import { logger } from '../../common/utils/logger';
 import { isHttpOrWsRequest, isHttpRequest, getHost } from '../../common/utils/url';
 
@@ -26,7 +26,7 @@ import { type ConfigurationMV3 } from './configuration';
 
 const ASYNC_LOAD_CHINK_SIZE = 5000;
 
-type EngineConfig = Pick<ConfigurationMV3, 'userrules' | 'quickFixesRules' | 'verbose'> & {
+type EngineConfig = Pick<ConfigurationMV3, 'userrules' | 'verbose'> & {
     filters: IFilter[];
 };
 
@@ -69,7 +69,6 @@ export class EngineApi {
         const {
             filters,
             userrules,
-            quickFixesRules,
             verbose,
         } = config;
 
@@ -108,20 +107,6 @@ export class EngineApi {
                     false,
                     false,
                     userrules.sourceMap,
-                ),
-            );
-        }
-
-        if (quickFixesRules.filterList.length > 0) {
-            // Note: rules are already converted at the extension side
-            lists.push(
-                new BufferRuleList(
-                    QUICK_FIXES_FILTER_ID,
-                    quickFixesRules.filterList,
-                    false,
-                    false,
-                    false,
-                    quickFixesRules.sourceMap,
                 ),
             );
         }